@@ -83,81 +83,81 @@
     "packaging": "jar"
   },
   "dependencies": {
-    "@angular/common": "16.0.5",
-    "@angular/compiler": "16.0.5",
-    "@angular/core": "16.0.5",
-    "@angular/forms": "16.0.5",
-    "@angular/localize": "16.0.5",
-    "@angular/platform-browser": "16.0.5",
-    "@angular/platform-browser-dynamic": "16.0.5",
-    "@angular/router": "16.0.5",
+    "@angular/common": "16.1.4",
+    "@angular/compiler": "16.1.4",
+    "@angular/core": "16.1.4",
+    "@angular/forms": "16.1.4",
+    "@angular/localize": "16.1.4",
+    "@angular/platform-browser": "16.1.4",
+    "@angular/platform-browser-dynamic": "16.1.4",
+    "@angular/router": "16.1.4",
     "@fortawesome/angular-fontawesome": "0.13.0",
     "@fortawesome/fontawesome-svg-core": "6.4.0",
     "@fortawesome/free-solid-svg-icons": "6.4.0",
-    "@ng-bootstrap/ng-bootstrap": "15.0.1",
+    "@ng-bootstrap/ng-bootstrap": "15.1.0",
     "@ngx-translate/core": "15.0.0",
     "@ngx-translate/http-loader": "8.0.0",
     "@popperjs/core": "2.11.8",
     "bootstrap": "5.3.0",
-    "dayjs": "1.11.8",
+    "dayjs": "1.11.9",
     "ngx-infinite-scroll": "16.0.0",
     "rxjs": "7.8.1",
-    "tslib": "2.5.3",
-    "zone.js": "0.13.0"
+    "tslib": "2.6.0",
+    "zone.js": "0.13.1"
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "16.0.0",
     "@angular-builders/jest": "16.0.0",
-    "@angular-devkit/build-angular": "16.0.5",
+    "@angular-devkit/build-angular": "16.1.4",
     "@angular-eslint/eslint-plugin": "16.0.2",
-    "@angular/cli": "16.0.5",
-    "@angular/compiler-cli": "16.0.5",
-    "@angular/service-worker": "16.0.5",
-    "@types/jest": "29.5.1",
-    "@types/node": "16.18.34",
-    "@typescript-eslint/eslint-plugin": "5.59.9",
-    "@typescript-eslint/parser": "5.59.9",
+    "@angular/cli": "16.1.4",
+    "@angular/compiler-cli": "16.1.4",
+    "@angular/service-worker": "16.1.4",
+    "@types/jest": "29.5.3",
+    "@types/node": "18.16.19",
+    "@typescript-eslint/eslint-plugin": "5.61.0",
+    "@typescript-eslint/parser": "5.61.0",
     "browser-sync": "2.29.3",
     "browser-sync-webpack-plugin": "2.3.0",
-    "concurrently": "8.1.0",
+    "concurrently": "8.2.0",
     "copy-webpack-plugin": "11.0.0",
-    "cypress": "12.14.0",
+    "cypress": "12.17.1",
     "cypress-audit": "1.1.0",
-    "eslint": "8.42.0",
+    "eslint": "8.44.0",
     "eslint-config-prettier": "8.8.0",
     "eslint-plugin-cypress": "2.13.3",
     "eslint-webpack-plugin": "4.0.1",
     "folder-hash": "4.0.4",
     "generator-jhipster": "7.9.3",
     "husky": "8.0.3",
-    "jest": "29.5.0",
+    "jest": "29.6.1",
     "jest-date-mock": "1.0.8",
-    "jest-environment-jsdom": "29.5.0",
+    "jest-environment-jsdom": "29.6.1",
     "jest-junit": "16.0.0",
     "jest-preset-angular": "13.1.1",
     "jest-sonar": "0.2.16",
-    "lighthouse": "10.2.0",
-    "lint-staged": "13.2.2",
+    "lighthouse": "10.4.0",
+    "lint-staged": "13.2.3",
     "merge-jsons-webpack-plugin": "2.0.1",
     "prettier": "2.8.8",
-    "prettier-plugin-java": "2.1.0",
-    "prettier-plugin-packagejson": "2.4.3",
+    "prettier-plugin-java": "2.2.0",
+    "prettier-plugin-packagejson": "2.4.5",
     "rimraf": "5.0.1",
-    "swagger-ui-dist": "4.19.0",
-    "ts-jest": "29.1.0",
-    "typescript": "5.0.4",
+    "swagger-ui-dist": "5.1.0",
+    "ts-jest": "29.1.1",
+    "typescript": "5.1.6",
     "wait-on": "7.0.1",
     "webpack-bundle-analyzer": "4.9.0",
     "webpack-merge": "5.9.0",
     "webpack-notifier": "1.15.0"
   },
   "engines": {
-    "node": ">=18.16.0"
+    "node": ">=18.16.1"
   },
   "cacheDirectories": [
     "node_modules"
   ],
   "overrides": {
-    "webpack": "5.86.0"
+    "webpack": "5.88.1"
   }
 }
