@@ -1,36 +1,33 @@
-import { Injectable, isDevMode } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
-import { Observable } from 'rxjs';
+import { inject, isDevMode } from '@angular/core';
+import { ActivatedRouteSnapshot, CanActivateFn, Router, RouterStateSnapshot } from '@angular/router';
 import { map } from 'rxjs/operators';
 
 import { AccountService } from 'app/core/auth/account.service';
 import { StateStorageService } from './state-storage.service';
 
-@Injectable({ providedIn: 'root' })
-export class UserRouteAccessService implements CanActivate {
-  constructor(private router: Router, private accountService: AccountService, private stateStorageService: StateStorageService) {}
+export const UserRouteAccessService: CanActivateFn = (next: ActivatedRouteSnapshot, state: RouterStateSnapshot) => {
+  const accountService = inject(AccountService);
+  const router = inject(Router);
+  const stateStorageService = inject(StateStorageService);
+  return accountService.identity().pipe(
+    map(account => {
+      if (account) {
+        const authorities = next.data['authorities'];
 
-  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
-    return this.accountService.identity().pipe(
-      map(account => {
-        if (account) {
-          const authorities = route.data['authorities'];
-
-          if (!authorities || authorities.length === 0 || this.accountService.hasAnyAuthority(authorities)) {
-            return true;
-          }
-
-          if (isDevMode()) {
-            console.error('User has not any of required authorities: ', authorities);
-          }
-          this.router.navigate(['accessdenied']);
-          return false;
+        if (!authorities || authorities.length === 0 || accountService.hasAnyAuthority(authorities)) {
+          return true;
         }
 
-        this.stateStorageService.storeUrl(state.url);
-        this.router.navigate(['/login']);
+        if (isDevMode()) {
+          console.error('User has not any of required authorities: ', authorities);
+        }
+        router.navigate(['accessdenied']);
         return false;
-      })
-    );
-  }
-}
+      }
+
+      stateStorageService.storeUrl(state.url);
+      router.navigate(['/login']);
+      return false;
+    })
+  );
+};
