@@ -1,22 +1,22 @@
 import { ILabel, NewLabel } from './label.model';
 
 export const sampleWithRequiredData: ILabel = {
-  id: 10119,
+  id: 3315,
   label: 'Designer',
 };
 
 export const sampleWithPartialData: ILabel = {
-  id: 87689,
-  label: 'grey Luciano Southwest',
+  id: 28734,
+  label: 'Tools purple Lead',
 };
 
 export const sampleWithFullData: ILabel = {
-  id: 58539,
-  label: 'productivity Director Security',
+  id: 12754,
+  label: 'Bicycle',
 };
 
 export const sampleWithNewData: NewLabel = {
-  label: 'Chief resource',
+  label: 'magenta Progressive',
   id: null,
 };
 
